NHSO-13570: Document Download xamarin tests

+++ xamarin/NHSOnline.App.iOS/DependencyServices/FileSystem/IosFileSystem.cs
@@ -23,9 +23,20 @@ namespace NHSOnline.App.iOS.DependencyServices.FileSystem
 
         public async void PresentFileActionController(DownloadRequest downloadRequest)
         {
-            await File.WriteAllBytesAsync(
-                downloadRequest.FileCachePath,
-                Convert.FromBase64String(downloadRequest.Base64Data)).ConfigureAwait(true);
+            if (!File.Exists(downloadRequest.FileCachePath))
+            {
+                try
+                {
+                    await File.WriteAllBytesAsync(
+                        downloadRequest.FileCachePath,
+                        Convert.FromBase64String(downloadRequest.Base64Data)).ConfigureAwait(true);
+                }
+                catch (Exception e)
+                {
+                    Logger.LogError("Failed to write data to cache directory", e);
+                    return;
+                }
+            }
 
             ShareFile shareFile = new ShareFile(downloadRequest.FileCachePath);
